If same fullpath ArchiveJob already exists, not upload the file

+++ app/models/glacier.rb
@@ -4,6 +4,8 @@ class Glacier
   end
 
   def archive_create
+    return nil if ArchiveJob.exists?(fullpath: @path)
+
     archive = vault.archives.create body: File.open(@path), description: File.basename(@path), :multipart_chunk_size => 1024*1024
     ArchiveJob.create!(jid: archive.id, description: File.basename(@path), fullpath: @path)
   end
